pyproject.toml
[tool.poetry]
name = "python-developer"
version = "0.1.0"
description = ""
authors = ["Feng Lu <feng@instaffo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.9"
elasticsearch = "^8.17.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
